+++ internal/docs/bloblang.go
@@ -2,7 +2,6 @@ package docs
 
 import (
 	"bytes"
-	"fmt"
 	"strings"
 	"text/template"
 
@@ -15,7 +14,7 @@ import (
 func LintBloblangMapping(ctx LintContext, line, col int, v interface{}) []Lint {
 	str, ok := v.(string)
 	if !ok {
-		return []Lint{NewLintError(line, fmt.Sprintf("expected string value, got %T", v))}
+		return nil
 	}
 	if str == "" {
 		return nil
@@ -38,7 +37,7 @@ func LintBloblangMapping(ctx LintContext, line, col int, v interface{}) []Lint {
 func LintBloblangField(ctx LintContext, line, col int, v interface{}) []Lint {
 	str, ok := v.(string)
 	if !ok {
-		return []Lint{NewLintWarning(line, fmt.Sprintf("expected string value, got %T", v))}
+		return nil
 	}
 	if str == "" {
 		return nil