+++ lib/processor/grok_test.go
@@ -125,14 +125,9 @@ func TestGrok(t *testing.T) {
 }
 
 func TestGrokFileImports(t *testing.T) {
-	tmpDir, err := os.MkdirTemp("", "grok_test")
-	require.NoError(t, err)
+	tmpDir := t.TempDir()
 
-	t.Cleanup(func() {
-		os.RemoveAll(tmpDir)
-	})
-
-	err = os.WriteFile(filepath.Join(tmpDir, "foos"), []byte(`
+	err := os.WriteFile(filepath.Join(tmpDir, "foos"), []byte(`
 FOOFLAT %{WORD:first} %{WORD:second} %{WORD:third}
 FOONESTED %{INT:nested.first:int} %{WORD:nested.second} %{WORD:nested.third}
 `), 0o777)